Moved WalletDriverError class in a separate file.
